added workflows, changed file structures and implememnted cache_sate_var detection
