+++ src/loader.rs
@@ -2,7 +2,6 @@ use reqwest;
 use reqwest::Error;
 use serde_json::{Map, Value};
 use std::fs;
-use crate::config;
 use crate::optimizor;
 
 fn read_sol_file(file_path: &str) -> Result<String, std::io::Error> {
@@ -12,17 +11,17 @@ fn read_sol_file(file_path: &str) -> Result<String, std::io::Error> {
     Ok(content)
 }
 
-async fn read_from_github(path: &str) -> Result<String, Error>  {
-    // Send an HTTP GET request to fetch the raw content
-    //Path should be in this format: https://raw.githubusercontent.com/user/repo/master/contract.sol
-    let client = reqwest::Client::new();
-    let response = client.get(path).send().await?;
-    let contract_content = response.text().await?;
-    println!("{}", contract_content);
+// async fn read_from_github(path: &str) -> Result<String, Error>  {
+//     // Send an HTTP GET request to fetch the raw content
+//     //Path should be in this format: https://raw.githubusercontent.com/user/repo/master/contract.sol
+//     let client = reqwest::Client::new();
+//     let response = client.get(path).send().await?;
+//     let contract_content = response.text().await?;
+//     println!("{}", contract_content);
 
-     Ok(contract_content)
+//      Ok(contract_content)
 
-}
+// }
 
 
 pub async fn loader() {
@@ -49,21 +48,20 @@ pub async fn loader() {
 
     optimizor::cache_state_variables::cache_state_variables(&mut gas_inefficiencies, 0);
     optimizor::write_zero_to_storage::write_zero_to_storage(&mut gas_inefficiencies, 0);
-    optimizor::gas_tricks::bytes32(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::openzepplin(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::safemath(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::token(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::uint_incur_overhead(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::check_constructor_absence(&contract);
-    optimizor::gas_tricks::use_named_retunrs(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::uint_incur_overhead(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::mapping_instead_array(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::uint256_instead_bool(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::require_double_logic(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::revert_32(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::do_while(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::priv_constants_immut(&contract, &mut gas_inefficiencies);
-    optimizor::gas_tricks::emit_loops(&contract, &mut gas_inefficiencies);
+    optimizor::bytes32::bytes32(&contract, &mut gas_inefficiencies);
+    optimizor::openzeppelin::openzepplin(&contract, &mut gas_inefficiencies);
+    optimizor::safemath::safemath(&contract, &mut gas_inefficiencies);
+    optimizor::token::token(&contract, &mut gas_inefficiencies);
+    optimizor::uint_incur_overhead::uint_incur_overhead(&contract, &mut gas_inefficiencies);
+    optimizor::constructor_check_absence::check_constructor_absence(&contract);
+    optimizor::use_named_returns::use_named_retunrs(&contract, &mut gas_inefficiencies);
+    optimizor::mapping_instead_array::mapping_instead_array(&contract, &mut gas_inefficiencies);
+    optimizor::uint256_instead_bool::uint256_instead_bool(&contract, &mut gas_inefficiencies);
+    optimizor::require_double_logic::require_double_logic(&contract, &mut gas_inefficiencies);
+    optimizor::revert_32::revert_32(&contract, &mut gas_inefficiencies);
+    optimizor::do_while::do_while(&contract, &mut gas_inefficiencies);
+    optimizor::priv_constants_immut::priv_constants_immut(&contract, &mut gas_inefficiencies);
+    optimizor::emit_loops::emit_loops(&contract, &mut gas_inefficiencies);
 
     // Convert the gas inefficiencies to JSON
     let gas_inefficiencies_json =
