+++ src/optimizor/cache_state_variables.rs
@@ -1,6 +1,6 @@
 use crate::utils;
 use regex::Regex;
-use serde_json::json;
+// use serde_json::json;
 use serde_json::Map;
 use serde_json::Value;
 use std::fs;
@@ -9,11 +9,9 @@ pub fn cache_state_variables(
     gas_inefficiencies: &mut Map<String, serde_json::Value>,
     mut _prev: usize,
 ) {
-    let arr_str: Vec<&str>;
     let ast_json = fs::read_to_string("src/optimizor/ast.json").expect("Failed to read");
     let ast: Value = serde_json::from_str(&ast_json).expect("Failed to deserialize");
-    let mut _name = "";
-    utils::get_all_state_var::get_all_state_vars();
+
     if let Some(nodes) = ast.get("nodes").and_then(Value::as_array) {
         for node in nodes {
             if let Some(node_type) = node.get("nodeType").and_then(Value::as_str) {
@@ -34,38 +32,105 @@ pub fn cache_state_variables(
                                                 if expression
                                                     .get("operator")
                                                     .and_then(Value::as_str)
+                                                    == Some("=")
+                                                {
+                                                    if !expression
+                                                        .get("leftHandSide")
+                                                        .unwrap()
+                                                        .get("name")
+                                                        .is_none()
+                                                    {
+                                                        let name = expression
+                                                            .get("leftHandSide")
+                                                            .unwrap()
+                                                            .get("name")
+                                                            .and_then(Value::as_str)
+                                                            .unwrap()
+                                                            .to_owned();
+
+                                                        let maps = utils::get_all_state_var::get_all_state_vars();
+                                                        let _name_bool: Option<bool> = maps
+                                                            .get(&name)
+                                                            .and_then(Value::as_bool);
+
+                                                        _prev = get_line_number_zero(&name, _prev);
+                                                        let inefficiency_id =
+                                                            format!("line_{}", _prev);
+
+                                                        utils::update_gas_ineffiency_map::update_gas_inefficency_map(inefficiency_id, gas_inefficiencies, "cache state variables".to_owned());
+                                                    } else if let Some(name) = expression
+                                                        .get("rightHandSide")
+                                                        .unwrap()
+                                                        .get("name")
+                                                        .and_then(Value::as_str)
+                                                        .map(String::from)
+                                                    {
+                                                        let maps = utils::get_all_state_var::get_all_state_vars();
+                                                        let _name_bool: Option<bool> = maps
+                                                            .get(&name)
+                                                            .and_then(Value::as_bool);
+
+                                                        println!("{}", expression);
+
+                                                        _prev = get_line_number_zero(&name, _prev);
+                                                        let inefficiency_id =
+                                                            format!("line_{}", _prev);
+
+                                                        utils::update_gas_ineffiency_map::update_gas_inefficency_map(inefficiency_id, gas_inefficiencies,"cache state variables".to_owned());
+                                                    }
+                                                } else if expression
+                                                    .get("operator")
+                                                    .and_then(Value::as_str)
                                                     != Some("=")
                                                 {
-                                                    if let Some(left_hand_side) =
-                                                        expression.get("leftHandSide")
+                                                    if let Some(arguments) = expression
+                                                        .get("body")
+                                                        .and_then(|b| b.get("arguments"))
+                                                        .and_then(Value::as_array)
                                                     {
-                                                        _name = left_hand_side
-                                                            .get("name")
-                                                            .unwrap()
-                                                            .as_str()
-                                                            .unwrap_or("");
-                                                        let maps = utils::get_all_state_var::get_all_state_vars();
-                                                        let name_bool = maps.get(_name).unwrap_or(
-                                                            &serde_json::to_value(false).unwrap(),
-                                                        );
-
-                                                        if let Some(right_hand_side) =
-                                                            expression.get("rightHandSide")
-                                                        {
-                                                            if let Some(right_hand_side_name) =
-                                                                right_hand_side.get("name")
+                                                        for argument in arguments {
+                                                            if let Some(left_E) =
+                                                                argument.get("leftExpression")
                                                             {
-                                                                if let Some(is_pure) =
-                                                                    right_hand_side.get("isPure")
-                                                                {
-                                                                    if is_pure == false {
-                                                                        if _name
-                                                                            == right_hand_side_name
-                                                                        {
-                                                                            /////**CACHE VARIABLE */
-                                                                        }
-                                                                    }
-                                                                }
+                                                                let name = argument
+                                                                    .get("leftExpression")
+                                                                    .unwrap()
+                                                                    .get("name")
+                                                                    .and_then(Value::as_str)
+                                                                    .unwrap()
+                                                                    .to_owned();
+
+                                                                let maps = utils::get_all_state_var::get_all_state_vars();
+                                                                let _name_bool: Option<bool> = maps
+                                                                    .get(&name)
+                                                                    .and_then(Value::as_bool);
+
+                                                                _prev = get_line_number_zero(
+                                                                    &name, _prev,
+                                                                );
+                                                                let inefficiency_id =
+                                                                    format!("line_{}", _prev);
+
+                                                                utils::update_gas_ineffiency_map::update_gas_inefficency_map(inefficiency_id, gas_inefficiencies, "cache state variables".to_owned());
+                                                            } else if let Some(name) = argument
+                                                                .get("rightExpression")
+                                                                .unwrap()
+                                                                .get("name")
+                                                                .and_then(Value::as_str)
+                                                                .map(String::from)
+                                                            {
+                                                                let maps = utils::get_all_state_var::get_all_state_vars();
+                                                                let _name_bool: Option<bool> = maps
+                                                                    .get(&name)
+                                                                    .and_then(Value::as_bool);
+
+                                                                _prev = get_line_number_zero(
+                                                                    &name, _prev,
+                                                                );
+                                                                let inefficiency_id =
+                                                                    format!("line_{}", _prev);
+
+                                                                utils::update_gas_ineffiency_map::update_gas_inefficency_map(inefficiency_id, gas_inefficiencies, "cache state variables".to_owned());
                                                             }
                                                         }
                                                     }
@@ -84,6 +149,9 @@ pub fn cache_state_variables(
 }
 
 fn get_line_number_zero(src: &str, mut _prev: usize) -> usize {
+    //get number of times
+    let mut times = 0;
+    println!("{}", "erere");
     // Read the source file as a string
     let contract = fs::read_to_string("src/contract.sol").expect("Failed to read");
 
@@ -91,16 +159,22 @@ fn get_line_number_zero(src: &str, mut _prev: usize) -> usize {
     let lines: Vec<&str> = contract.lines().collect();
 
     // Format the string with " = 0" at the end
-    let strss = format!(r"{} = 0", src);
+    let strss = format!(r"{}", src);
+
+    // Compile the regex pattern for any function declaration
+    let any_function_declaration_regex = Regex::new(r"function\s+\w*\s*\(").unwrap();
 
     // Compile the regex pattern
     let variable_declaration_regex = Regex::new(&strss).unwrap();
 
     for (line_number, line) in lines.iter().enumerate() {
-        if let Some(_capture) = variable_declaration_regex.captures(line) {
-            if line_number > _prev {
-                _prev = line_number + 1;
-                break;
+        if let Some(_captures) = any_function_declaration_regex.captures(line) {
+            if let Some(_capture) = variable_declaration_regex.captures(line) {
+                if line_number > _prev && times > 1 {
+                    _prev = line_number + 1;
+                    break;
+                }
+                times = times + 1;
             }
         }
     }
