+++ src/optimizor/mod.rs
@@ -1,4 +1,18 @@
 pub mod ast;
-pub mod gas_tricks;
+pub mod token;
+pub mod bytes32;
+pub mod safemath;
+pub mod openzeppelin;
+pub mod revert_32;
+pub mod do_while;
+pub mod emit_loops;
 pub mod write_zero_to_storage;
 pub mod cache_state_variables;
+pub mod uint256_instead_bool;
+pub mod require_double_logic;
+pub mod mapping_instead_array;
+pub mod priv_constants_immut;
+pub mod pack_related_vars;
+pub mod use_named_returns;
+pub mod uint_incur_overhead;
+pub mod constructor_check_absence;
